🔥 remove useProjectLeadInvitations hook
